sistema-corporativo/backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
python-multipart
pydantic
pydantic-settings
email-validator
